Add files via upload
Hello everyone, I have completed a project for CodeBasics Challenge No. 6
ABOUT PROJET : CodeX is a German beverage company that is aiming to make its mark in the Indian market. A few months ago, they launched their energy drink in 10 cities in India.
Their Marketing team is responsible for increasing brand awareness, market share, and product development. They conducted a survey in those 10 cities and received results from 10k respondents my task is to convert these survey results to meaningful insights which the team can use to drive actions.

I chose an SQL platform for this project because showing graphs would not be a great approach. This was not a scenario involving sales data where we would need to show monthly trends and compare them with our targets. The data we had was based on a survey, so we needed to find answers based on the maximum number of responses. That's why SQL was a great choice, in my opinion

Regarding the data, it was in the form of Excel sheets containing the survey response table, respondents' information, and information about the city where the product was launched. First, I imported the data into SQL, and with the help of queries, I found the answers to important questions

Below are the insights in the screenshot from where I found the answers to some questions using queries. Based on those answers, here are some suggestions for CodeX:
 
1. Insights tell us that most people think the ideal price should be in the range of 55-99 INR.
2. From insights, we see that the age group of 19-30 drinks energy drinks daily or 2-3 times a week, typically during sports and exercise. Therefore, we can make a popular young athlete the brand ambassador for our drink.
3. The male audience prefers energy drinks, so codex should target males aged between 19-30.
4. Some people like caffeine or vitamins in their drinks, so codex can launch new flavors to cater to their preferences. This also includes diet flavors to attract health-conscious people.
5. Stores are the common platform where people buy most of our drinks, so codex should focus on them and advertise there with hoardings.
6. Most people give our drink 3 stars for taste, so codex should work on improving the taste to compete with big players in the market like Cola-Coca and Bepsi.
7. codex should focus more on online ads, as most people learn about their drink through online ads.